requirements.txt
selenium==4.8.2
webdriver-manager==3.8.5
pytest==7.2.2
allure-pytest==2.13.0
allure-python-commons==2.13.0